login_app/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Werkzeug==2.3.7
supabase==2.3.4
python-dotenv==1.0.0
python-jose==3.3.0
